refactor: extract chat turn contracts
