+++ docs/local/2026-08-15-frontend-modularization-and-slimming-plan.md
@@ -1,7 +1,7 @@
 # Tinybot 前端模块化与瘦身计划
 
 - 日期：2026-08-15
-- 状态：实施中；Phase 7 已按职责清晰度收口，低风险瘦身、TinyOS/Memory/Settings/Tools 加载 seam、首批路由 CSS、Shell route surface、Workspace、Tools/Plugins、Settings 与 native event adapter、TinyOS 循环清理已完成
+- 状态：实施中；Phase 7 已按职责清晰度收口，Phase 3 的稳定 Chat contracts 已完成独立，backend payload 校验与 timeline 投影仍待继续收窄
 - 基线提交：`c18d0bae refactor: extract chat context usage`
 - 范围：`src/react-workbench`、被桌面前端直接使用的 `src/app-core`、前端依赖和分析工具
 - 本地约束：本文位于被忽略的 `docs/local/`，只作为本地实施依据，不推送到 GitHub
@@ -76,6 +76,10 @@
 - `defaultServices.ts` 从约 854 行继续降至约 692 行；新增 event bridge interface 测试，覆盖四类 listener、workspace child discovery、terminal event 去重、form 状态、browser/host 错误和 dispatch 状态一致性；
 - 完整分析通过 87 个测试文件、572 个测试、类型检查、源码分析、构建与 bundle 门禁；ESLint finding 保持 43，生产循环保持 0，初始资源 gzip 为 486,029 B，比上轮增加 253 B。
 - Phase 7 在此收口：剩余 `defaultServices.ts` 主体是 controller、SessionStore、ChatStore 与 command dispatch 的同一条会话编排链，继续拆分会制造 callback 拼装并与 Phase 4 重叠；Memory/ProjectGroup 则只是薄 native 转发，不为达到 `<400` 行目标单独提取。
+- 新建 `chatTurnContracts.ts`，集中稳定的 UI turn、canonical item、runtime state、artifact、usage、form/plan 与 delegated agent contracts；文件只含 type import/type export，不执行 payload 校验或 UI 投影；
+- 所有类型调用方已直接迁移到 contracts 模块，`chatTurnModel.ts` 不保留 compatibility re-export，类型消费者不再依赖约 50 KB 的投影实现源码；
+- `chatTurnModel.ts` 从约 1,375 行降至 1,060 行，只保留 artifact 投影、canonical payload 规范化、timeline 到 UI turn 投影和安全预览；Phase 3 下一切片应继续分开 payload 边界与 projection interface；
+- Chat contracts 切片完整分析通过：87 个测试文件、572 个测试、TypeScript、ESLint、源码分析、生产构建和 bundle 门禁全部成功；ESLint finding 保持 43，生产循环和不可达候选保持 0，初始资源 gzip 保持 486,029 B。
 
 Settings/TinyOS 剩余路由级 CSS、Chat/Settings/TinyOS 模块深化与静态分析债务仍待后续阶段实施。
 
@@ -344,6 +348,8 @@ createDesktopAppServices           # single external interface remains
    - backend timeline 到 UI turn 的投影与安全预览；
 4. 为投影提供单一 interface，例如 `projectBackendTimeline(input)`，内部 helper 不导出；
 5. 迁移调用方到准确模块，临时 compatibility re-export 只允许在同一阶段内存在，阶段结束前删除。
+
+当前状态：稳定 contracts 拆分已完成，调用方已迁移且没有 compatibility re-export。payload 校验/规范化与 timeline projection 仍位于 `chatTurnModel.ts`，因此 Phase 3 尚未完成；下一步优先定义单一 projection interface，再移动其私有实现与对应测试，避免为了减行数复制通用 helper。
 
 退出条件：
