+++ src/react-workbench/chat/PatchDiffCard.test.tsx
@@ -3,7 +3,7 @@
 import { cleanup, render, screen, waitFor, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { afterEach, describe, expect, it, vi } from "vitest";
-import type { ToolCallState } from "../../app-core/chat/chatTurnModel";
+import type { ToolCallState } from "../../app-core/chat/chatTurnContracts";
 import { PatchDiffCard, patchChangeSetFromToolResult } from "./PatchDiffCard";
 
 const applyPatchToolCall: ToolCallState = {
