+++ src/react-workbench/chat/test/timelineFixtures.ts
@@ -1,5 +1,5 @@
 import type { ChatTimelineSnapshot } from "../../../app-core/chat/agentTimelineModel";
-import type { ChatStep, ChatTurn } from "../../../app-core/chat/chatTurnModel";
+import type { ChatStep, ChatTurn } from "../../../app-core/chat/chatTurnContracts";
 import type { ReactChatMessage } from "../messageActions";
 
 export function timelineFromReactMessages(
